wrapped css styles in stack.css
